reducers: extract superglueReducer to its own file
